requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
requests==2.22.0
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.6
